next steps copy just for ap
