+++ apps/patient/src/components/order-details/OrderDetailsModal.tsx
@@ -1,20 +1,21 @@
 import {
-  Container,
-  VStack,
-  Text,
-  HStack,
-  Modal,
-  ModalOverlay,
-  ModalContent,
-  ModalCloseButton,
-  ModalHeader,
-  ModalBody,
   Box,
-  Icon
+  Container,
+  HStack,
+  Icon,
+  Modal,
+  ModalBody,
+  ModalCloseButton,
+  ModalContent,
+  ModalHeader,
+  ModalOverlay,
+  Text,
+  VStack
 } from '@chakra-ui/react';
 import dayjs from 'dayjs';
 import { ReactNode } from 'react';
 import { MdOutlineLocalPharmacy } from 'react-icons/md';
+import { renderWithLinks } from '../../utils/text';
 
 export interface PrescriptionData {
   rxName: string;
@@ -26,14 +27,20 @@ export interface PrescriptionData {
 
 export interface OrderDetailsProps {
   pharmacyName: string;
+  pharmacyId?: string;
   pharmacyLogo?: ReactNode;
-
   prescriptions: PrescriptionData[];
 }
+
 export interface OrderDetailsModalProps {
   isOpen: boolean;
   onClose: () => void;
 }
+
+const NEXT_STEPS_BY_PHARMACY: Record<string, string> = {
+  [import.meta.env.VITE_AMAZON_PHARMACY_ID as string]:
+    'Amazon Pharmacy will text you shortly — no action needed.\n\nNo text? Log in at [amazon.com/pharmacy](https://amazon.com/pharmacy) to view your prescription or create an account.\n\nStill having trouble? Contact support. You can also switch pharmacies if needed.'
+};
 
 const Row = ({ k, value }: { k: string; value: ReactNode }) => {
   return (
@@ -70,6 +77,7 @@ const defaultIcon = (
 );
 
 export const OrderDetailsModal = (props: OrderDetailsProps & OrderDetailsModalProps) => {
+  const nextSteps = props.pharmacyId ? NEXT_STEPS_BY_PHARMACY[props.pharmacyId] : undefined;
   const handleClose = () => {
     props.onClose();
   };
@@ -110,6 +118,16 @@ export const OrderDetailsModal = (props: OrderDetailsProps & OrderDetailsModalPr
                   <PrescriptionBlock key={`${p.rxName}-${i}`} rx={p} />
                 ))}
               </VStack>
+              {nextSteps && (
+                <VStack alignItems="stretch" spacing={2} w="full">
+                  <Text fontWeight="bold" fontSize="lg">
+                    Next Steps
+                  </Text>
+                  <Box bgColor="blue.50" borderRadius="xl" p={4}>
+                    <Text whiteSpace="pre-wrap">{renderWithLinks(nextSteps)}</Text>
+                  </Box>
+                </VStack>
+              )}
             </VStack>
           </Container>
         </ModalBody>
